requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptography==1.7.2
Flask==0.12.2
Flask-Cors==3.0.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
pyasn1==0.3.7
pycparser==2.18
PyJWT==1.5.3
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
vine==1.1.4
Werkzeug==0.12.2
